requirements.txt
# Python dependencies for Person Tracking System

# Intel RealSense Camera
pyrealsense2>=2.50.0

# Computer Vision
opencv-python>=4.8.0
numpy>=1.24.0

# Deep Learning (Jetson optimized)
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0

# Google Gemini APIs (prefer google-genai, fallback to google-generativeai)
google-genai>=0.3.0
google-generativeai>=0.7.0
pillow>=10.0.0

# Firebase
firebase-admin>=6.0.0

# Screenshot capture
mss>=9.0.0

# Utilities
scipy>=1.10.0
matplotlib>=3.7.0
python-dotenv>=1.0.0
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code Quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
